% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costnp-package.R
\docType{package}
\name{costnp-package}
\alias{costnp-package}
\alias{_PACKAGE}
\title{costnp: Neyman-Pearson Cost Selection}
\description{
Given a cost-sensitive algorithm, outputs a cost and corresponding classifier with specific Neyman-Pearson guarantees.
}
\author{
\strong{Maintainer}: John Park \email{john.park8@gmail.com}

}
\keyword{internal}
